+++ apps/client/src/app/modules/list/list-controller.ts
@@ -225,6 +225,9 @@ export class ListController {
    */
   public static setDone(list: List, itemId: number | string, amount: number, excludeFinalItems = false, onlyFinalItems = false, setUsed = false, recipeId?: string, external = false, initialAddition = amount, newEtag = Date.now()): void {
     const item = ListController.getItemById(list, itemId, excludeFinalItems, onlyFinalItems, recipeId);
+    if (!item) {
+      return;
+    }
     const previousDone = item.amount_needed - MathTools.absoluteCeil((item.amount - item.done) / item.yield);
     if (setUsed) {
       // Save previous used amount